requirements.txt
screeninfo
simple_pid
qdarkstyle
pyqt5
matplotlib
pandas
pyserial
cobs
pyautogui
pyqtgraph
Pillow